Reorganized file structure for client
