+++ client/app/animal/[id]/page.js
@@ -1,5 +1,5 @@
 import Navbar from "@/app/components/layouts/Navbar";
-import ReadMoreById from "../../components/Miscellaneous/ReadMoreById";
+import ReadMoreById from "./components/ReadMoreById";
 import Footer from "@/app/components/layouts/Footer";
 
 const StrayDetailsPage = ({ params }) => {
