+++ client/app/reportlost/page.js
@@ -1,6 +1,6 @@
 'use client'
 
-import ReportLost from '../components/forms/ReportLost'
+import ReportLost from './components/ReportLost'
 import Navbar from '../components/layouts/Navbar'
 import Footer from '../components/layouts/Footer'
